review: extract shared inferfaces (NewsDataProvider and FetchNewsProviderOptions)
