+++ apps/crn-server/src/data-providers/types/index.ts
@@ -1,0 +1,1 @@
+export * from './news.data-provider.types';
