+++ apps/crn-server/test/mocks/news-data-provider.mock.ts
@@ -1,4 +1,4 @@
-import { NewsDataProvider } from '../../src/data-providers/news.data-provider';
+import { NewsDataProvider } from '../../src/data-providers/types';
 
 export const newsDataProviderMock: jest.Mocked<NewsDataProvider> = {
   fetch: jest.fn(),
